% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootresampling.R
\name{bootresampling}
\alias{bootresampling}
\title{A function for helping perform single-sample and k-sample bootstrap tests}
\usage{
bootresampling(x, stdx, stat, B, ...)
}
\arguments{
\item{x}{Symmetric matrix observations as a list of matrices, or list of list of matrices (see \code{\link[=as_flat]{as_flat()}}).}

\item{stdx}{A (multi)sample of matrices stored in flat form (\code{\link[=as_flat]{as_flat()}}) and standardized to satisfy the null hypothesis OR sampling weights for each matrix in \code{x}, in the same structure as \code{x}.}

\item{stat}{Function to compute the statistic.}

\item{B}{The number of bootstrap samples to use.}

\item{...}{Passed to \code{stat}}
}
\value{
A list of
\itemize{
\item \code{pval} the \code{p}-value from the test
\item \code{t0} the statistic for the observations \code{x}
\item \code{nullt} The statistic evaluated on the resamples
\item \code{stdx} The \code{stdx} passed into \code{bootresampling()}
\item \code{B} The number of resamples requested
\item \code{nullt_messages} Any error messages for the corresponding resample
}

The returned object has a bespoke class \code{TFORGE} for easy use of \code{print()}.
}
\description{
Designed for internal use.
Performs bootstrap resampling, calculation of a statistic and returns \code{p}-value of a hypothesis test.
Resampling is conducted as if the null hypothesis holds by using the \code{stdx} argument which contains either standardised matrices or weights.
}
\details{
The function \code{stat} is applied to \code{x} and all resamples. The statistic values are returned in the \code{t0} and \code{nullt} elements respectively.

Errors in evaluating the statistic \code{stat} on resamples are recorded in the \code{nullt_messages} and lead to \code{NA} values for the statistic.

The \code{p}-value is the fraction of non-\code{NA} resample statistic values that are greater than the \code{stat} applied to \code{x}.
\itemize{
\item \emph{include an option (\code{repair} say) to use common evals that aren't descending by reordering?}
}
}
