% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_evals.R
\name{cov_evals}
\alias{cov_evals}
\title{Eigenvalue covariance from eigenvectors and element-wise covariance}
\usage{
cov_evals(evecs, mcov)
}
\arguments{
\item{evecs}{Matrix of eigenvectors of the mean of \code{Y}. Eigenvectors as columns and ordered according to descending eigenvalue.}

\item{mcov}{Covariance of \code{vech(Y)}, where \code{Y} is the random matrix.}
}
\value{
A covariance matrix corresponding to the eigenvalues of the mean of \code{Y} in descending order.
}
\description{
For a random symmetric matrix \code{Y}, the covariance of the eigenvalues of \code{Y} is calculated from the covariance of the elements and the eigenvectors of the mean of \code{Y}.
This calculation is valid when the eigenvalues are distinct.
}
\details{
Uses \code{RcppArmadillo} for a slight increase in speed.
}
