% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedtrace.R
\name{has_fixedtrace}
\alias{has_fixedtrace}
\title{Check if the supplied sample(s) have fixed trace}
\usage{
has_fixedtrace(x, tolerance = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{A sample or multiple samples of matrices suitable for \code{\link[=as_flat]{as_flat()}}.}

\item{tolerance}{Tolerance on the relative difference, passed to \code{\link[=all.equal]{all.equal()}}}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Compares the trace of all the supplied matrices to check that they are equal.
}
