% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss1.R
\name{normalise_ss1}
\alias{normalise_ss1}
\title{Normalise so that Sum of Squared Eigenvalues is One}
\usage{
normalise_ss1(x)
}
\arguments{
\item{x}{A sample of matrices suitable for \code{\link[=as_fsm]{as_fsm()}}.}
}
\value{
A \code{TFORGE_fsm} object.
}
\description{
Scales symmetric tensors so that the square of the eigenvalues sum to one.
}
