% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedtrace.R
\name{normalise_trace}
\alias{normalise_trace}
\alias{normalize_trace}
\title{Scale symmetric matrices to have trace of one}
\usage{
normalise_trace(x)

normalize_trace(x)
}
\arguments{
\item{x}{A sample of matrices suitable for \code{\link[=as_fsm]{as_fsm()}}.}
}
\value{
A set of flattened symmetric matrices (i.e. \code{\link[=as_fsm]{TFORGE_fsm}} class).
}
\description{
Scales symmetric matrices by their trace, so that resulting matrices have a trace of one.
}
\details{
The method will create \code{Inf} values for tensors that have a trace of zero.
}
