% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicity.R
\name{test_multiplicity}
\alias{test_multiplicity}
\alias{stat_multiplicity}
\alias{standardise_multiplicity}
\title{Test eigenvalue multiplicity}
\usage{
test_multiplicity(x, mult, B = 1000)

stat_multiplicity(x, mult, evecs = NULL)

standardise_multiplicity(x, mult)
}
\arguments{
\item{x}{A single sample of matrices (passed to \code{\link[=as_fsm]{as_fsm()}}).}

\item{mult}{A vector specifying the eigenvalue multiplicity under the null hypothesis in descending order of eigenvalue size.}

\item{B}{The number of bootstrap samples}

\item{evecs}{For debugging only. Supply eigenvectors of population mean.}
}
\value{
\itemize{
\item \code{test_multiplicity()} returns a \code{TFORGE} object including the p-value of the test (slot \code{pval}) and the statistic for \code{x} (slot \code{t0}). See \code{\link[=bootresampling]{bootresampling()}}.
}

\itemize{
\item \code{stat_multiplicity()} returns a single value.
}

\itemize{
\item \code{standardise_multiplicity()} returns a sample of matrices stored in flattened form (a \code{TFORGE_fsm}).
}
}
\description{
Given a sample from a population, uses bootstrap resampling to test multiplicity hypotheses of the population mean's eigenvalues.
The test statistic is computed by \code{stat_multiplicity()}, which includes a uniformly random rotation of eigenvectors associated with each eigenvalue of multiplicity greater than 1.
The null hypothesis is the multiplicity of eigenvalues of the population mean.
Bootstrap resampling is conducted from the null hypothesis, which uses the original sample converted to satisfy the null hypothesis by \code{standardise_multiplicity()}.
}
\details{
This hypothesis test works on unconstrained symmetric matrices or matrices constrained to have fixed trace. It may work poorly on matrices with other constraints, on samples smaller than 15, or samples of multimodal populations.

Due to the random rotation of the eigenvectors, use \code{\link[=set.seed]{set.seed()}} before \code{stat_multiplicity()} or \code{test_multiplicity()} if you want the answer to be repeatable.
}
\examples{
x <- rsymm_norm(15, mean = diag(c(2, 1, 1, 0)))
test_multiplicity(x, mult = c(1, 2, 1))
}
