% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss1fixedtrace.R
\name{test_ss1fixedtrace}
\alias{test_ss1fixedtrace}
\alias{stat_ss1fixedtrace}
\title{Test eigenvalues when trace=0 and sum of square eigenvalues = 1}
\usage{
test_ss1fixedtrace(x, evals = NULL, B = 1000, maxit = 25)

stat_ss1fixedtrace(x, evals = NULL)
}
\arguments{
\item{x}{A single sample of symmetric matrices or multiple samples of symmetric matrices. See \code{\link[=as_flat]{as_flat()}}.}

\item{evals}{When \code{x} is a single sample, the null hypothesis is that the (extrinsic) mean of the population has eigenvalues equal to \code{evals}. For multiple samples \code{evals} must be omitted.}

\item{B}{Number of bootstrap samples. If \code{B = 'chisq'} then a chi-squared calibration is used instead.}

\item{maxit}{The maximum number of Newton steps allowed in empirical likelihood optimisation \insertCite{owen:2013}{TFORGE}.}
}
\value{
A \code{TFORGE} object (see \code{\link[=bootresampling]{bootresampling()}} or \code{\link[=chisq_calib]{chisq_calib()}}) with the eigenvalues of the null hypothesis in the \code{null_evals} attribute for \code{t0}.
}
\description{
For a single sample, test eigenvalues of the population mean.
For multiple samples, test for equality of the eigenvalues of the population means.
This function is for 3x3 symmetric matrices with trace of zero and sum of squared eigenvalues of one.
These constraints combine so that the space of possible sets of (ordered) eigenvalues is a circle.
The test statistic is calculated by \code{stat_ss1fixedtrace()}.
}
\details{
Test hypotheses described below.

The sum of squared eigenvalues constraint forces the set of eigenvalues to lie on a sphere and the trace constraint forces eigenvalues onto a plane.
Combined the constraints force eigenvalues onto a circle in 3D Euclidean space.
The test statistic accounts for these constraints by projecting eigenvalues onto a line tangential to this circle and orthogonal to the null-hypothesis eigenvalues.

Weighted bootstrap calibration is used (see 'Weighted Bootstrapping' below).

Eigenvalues must be distinct.
}
\section{Hypotheses}{
For a single sample the null hypothesis is that the population (extrinsic) mean has eigenvalues of \code{evals}; the alternative hypothesis is that the eigenvalues are not equal to \code{evals}.
For multiple samples, \code{evals} must be omitted and the null hypothesis is that the population (extrinsic) means have the same eigenvalues.
}

\section{Weighted Bootstrapping}{
This function uses a form of weighted bootstrapping called b-boostrapping \insertCite{hall1999in}{TFORGE}. An empirical distribution is defined by sampling weights for each observation in the original sample.
The sampling weights must be such that the (extrinsic) mean of the empirical distribution is \deqn{c\hat{Q} \Lambda \hat{Q}^\top,}
where \eqn{\hat{Q}} are the eigenvectors of the sample mean, \eqn{\Lambda} is a diagonal matrix of eigenvalues specified by either the null hypothesis (for single sample tests) or estimated as the common eigenvalues of multiple populations (for k-sample tests).
In some situations \eqn{c} is a free scalar to enable projection of the Euclidean mean to the extrinsic mean, otherwise \eqn{c=1}.
If no such sampling weights exist (i.e. the convex hull of the data does not contain \eqn{c\hat{Q} \Lambda \hat{Q}^\top}), then the test rejects with \code{pval=0} and a warning.

The sampling weights are also optimised to maximise empirical likelihood \insertCite{owen:2013}{TFORGE}.
}

\references{
\insertAllCited{}
}
