% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unconstrained.R
\name{test_unconstrained}
\alias{test_unconstrained}
\alias{stat_unconstrained}
\alias{standardise_specifiedevals}
\title{Pivotal bootstrap test of mean eigenvalues}
\usage{
test_unconstrained(x, evals = NULL, evecs = NULL, B = 1000)

stat_unconstrained(x, evals = NULL, evecs = NULL)

standardise_specifiedevals(x, evals)
}
\arguments{
\item{x}{A single sample of symmetric matrices or multiple samples of symmetric matrices. See \code{\link[=as_flat]{as_flat()}}.}

\item{evals}{When \code{x} is a single sample, the null hypothesis is that the population mean has eigenvalues \code{evals}.}

\item{evecs}{For a single sample, specify eigenvectors to test under the assumption that the population mean's eigenvectors are the columns of \code{evecs}. The order of these eigenvectors matters and should be such that eigenvalues are in descending order.}

\item{B}{Number of bootstrap samples.}
}
\value{
A \code{TFORGE} object with the eigenvalues of the null hypothesis in the \code{null_evals} attribute for \code{t0}. See \code{\link[=bootresampling]{bootresampling()}}.
}
\description{
For a single sample of symmetric matrices, test eigenvalues of the population mean.
For multilpe samples of symmetric matrices, test for equality of the eigenvalues of the population means.
Eigenvalues must be distinct.
}
\details{
For a single sample with \code{evecs} omitted, the null hypothesis is that the population mean has eigenvalues of \code{evals} without restriction on the eigenvectors. The alternative hypothesis here is that the eigenvalues are not \code{evals}.
If a \code{evecs} is supplied then both the null hypothesis and alternative hypothesis assume that the population mean has eigenvectors of \code{evecs}.

For multiple samples, \code{evals} and \code{evecs} must be omitted. The null hypothesis is that the population means of each sample have the same eigenvalues.

Bootstrap resampling is conducted from a population that satisfies the null hypothesis by transforming each sample in \code{x} with \code{standardise_specifiedevals()}.
The test statistic is calculated by \code{stat_unconstrained()}.
}
\examples{
test_unconstrained(rsymm_norm(15, diag(c(3,2,1))), evals = c(3, 2, 1), B = 100)
test_unconstrained(list(rsymm_norm(15, diag(c(3,2,1))),
                        rsymm_norm(15, diag(c(3,2,1)))), B = 100)
}
