% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_calibration.R
\name{chisq_calib}
\alias{chisq_calib}
\title{Internal: Chi Squared Calibration for Testing}
\usage{
chisq_calib(x, stat, df, ...)
}
\arguments{
\item{x}{Symmetric matrix observations as a list of matrices, or list of list of matrices (see \code{\link[=as_flat]{as_flat()}}).}

\item{stat}{Function to compute the statistic.}

\item{...}{Passed to \code{stat}}
}
\description{
Similar to \code{bootresampling()}, but uses a chi-squared distribution instead of an empirical distribution from bootstrapping.
}
