% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss1.R
\name{has_ss1}
\alias{has_ss1}
\title{Check whether the supplied sample(s) have equal sum of squared eigenvalues}
\usage{
has_ss1(x, tolerance = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{A sample or multiple samples of matrices suitable for \code{\link[=as_flat]{as_flat()}}.}

\item{tolerance}{Tolerance on the relative difference, passed to \code{\link[=all.equal]{all.equal()}}}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Compares whether the sum of the squared eigenvalues of the supplied symmetric matrices match each other using the property that the sum of the squared eigenvalues of \code{Y} equals the trace of \code{Y \%*\% Y}.
}
