% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedtrace.R
\name{project_trace}
\alias{project_trace}
\title{Project diagonal elements to have trace of zero}
\usage{
project_trace(x)
}
\arguments{
\item{x}{A sample of matrices suitable for \code{\link[=as_fsm]{as_fsm()}}.}
}
\description{
Projects the diagonal elements of symmetric matrices onto the plane through the origin and orthogonal to the vector \eqn{(1,1,1,....,1)^\top}.
The trace of the resulting symmetric matrices is zero.
}
