% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedtrace.R
\name{test_fixedtrace}
\alias{test_fixedtrace}
\alias{stat_fixedtrace}
\title{Test for eigenvalues when trace is fixed}
\usage{
test_fixedtrace(x, evals = NULL, B, maxit = 25)

stat_fixedtrace(x, evals = NULL)
}
\arguments{
\item{x}{A single sample of symmetric matrices or multiple samples of symmetric matrices. See \code{\link[=as_flat]{as_flat()}}.}

\item{evals}{When \code{x} is a single sample, the null hypothesis is that the population mean has eigenvalues \code{evals}.}

\item{B}{Number of bootstrap samples.}

\item{maxit}{The maximum number of Newton steps allowed in empirical likelihood optimisation \insertCite{owen:2013}{TFORGE}.}
}
\value{
A \code{TFORGE} object with the eigenvalues of the null hypothesis in the \code{null_evals} attribute for \code{t0}. See \code{\link[=bootresampling]{bootresampling()}}.
}
\description{
For a single sample of symmetric matrices with fixed trace, test eigenvalues of the population mean.
For multilpe samples of symmetric matrices with fixed trace, test for equality of the eigenvalues of the population means.
The test statistic is calculated by \code{stat_fixedtrace()}.
}
\details{
The fixed trace constraint forces the set of eigenvalues to lie in a plane.
The test statistic accounts for this constraint by using an orthonormal basis in the plane.
Bootstrap resampling is from an empirical distribution that satisfies the null hypothesis; for this test we use empirical likelihood \insertCite{owen:2013}{TFORGE} to find probability mass weights for each matrix in the original sample.

Eigenvalues must be distinct.

The test statistic is calculated by \code{stat_fixedtrace()}.
}
\section{Hypotheses}{
For a single sample the null hypothesis is that the population mean has eigenvalues of \code{evals}; the alternative hypothesis is that the eigenvalues are not \code{evals}.
For multiple samples, \code{evals} must be omitted and the null hypothesis is that the population means of each sample have the same eigenvalues.
}

\references{
\insertAllCited{}

\insertAllCited{}
}
