% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicity.R
\name{test_multiplicity}
\alias{test_multiplicity}
\alias{test_multiplicity_nonnegative}
\alias{stat_multiplicity}
\alias{standardise_multiplicity}
\title{Test eigenvalue multiplicity}
\usage{
test_multiplicity(x, mult, B = 1000, refbasis = "random")

test_multiplicity_nonnegative(
  x,
  mult,
  B = 1000,
  maxit = 25,
  refbasis = "random"
)

stat_multiplicity(x, mult, evecs = NULL, refbasis = "random")

standardise_multiplicity(x, mult)
}
\arguments{
\item{x}{A single sample of matrices (passed to \code{\link[=as_fsm]{as_fsm()}}).}

\item{mult}{A vector specifying the eigenvalue multiplicity under the null hypothesis in descending order of eigenvalue size.}

\item{B}{Number of bootstrap samples. If \code{B = 'chisq'} then a chi-squared calibration is used instead.}

\item{evecs}{For debugging only. Supply eigenvectors of population mean.}
}
\value{
A \code{TFORGE} object including the p-value of the test (slot \code{pval}) and the statistic for \code{x} (slot \code{t0}). See \code{\link[=bootresampling]{bootresampling()}}.
}
\description{
Given a sample from a population, uses bootstrap resampling to test multiplicity hypotheses of the population mean's eigenvalues.
The test statistic is computed by \code{stat_multiplicity()}, which includes a uniformly random rotation of eigenvectors associated with each eigenvalue with multiplicity greater than 1.
The null hypothesis is that the population mean has the specified the multiplicity of eigenvalues.
Bootstrap resampling is conducted from the null hypothesis, which uses the original sample converted to satisfy the null hypothesis by \code{standardise_multiplicity()}.
}
\details{
This hypothesis test works on unconstrained symmetric matrices or matrices constrained to have fixed trace. It may work poorly on matrices with other constraints, on samples smaller than 15, or samples of multimodal populations.

Due to the random rotation of the eigenvectors, use \code{\link[=set.seed]{set.seed()}} before \code{stat_multiplicity()} or \code{test_multiplicity()} if you want the answer to be repeatable.

\code{test_multiplicity_nonnegative()} uses weighted bootstrapping with empirical likelihood to create a population that satisfies the null hypothesis.
}
\examples{
x <- rsymm_norm(15, mean = diag(c(2, 1, 1, 0)))
test_multiplicity(x, mult = c(1, 2, 1))
}
