% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicity.R
\name{test_multiplicity}
\alias{test_multiplicity}
\alias{test_multiplicity_nonnegative}
\alias{stat_multiplicity}
\alias{standardise_multiplicity}
\title{Test eigenvalue multiplicity}
\usage{
test_multiplicity(x, mult, B = 1000, refbasis = "sample")

test_multiplicity_nonnegative(
  x,
  mult,
  B = 1000,
  maxit = 25,
  refbasis = "sample"
)

stat_multiplicity(x, mult, evecs = NULL, refbasis = "sample")

standardise_multiplicity(x, mult)
}
\arguments{
\item{x}{A single sample of matrices (passed to \code{\link[=as_fsm]{as_fsm()}}).}

\item{mult}{A vector specifying the eigenvalue multiplicity under the null hypothesis in descending order of eigenvalue size.}

\item{B}{Number of bootstrap samples. If \code{B = 'chisq'} then a chi-squared calibration is used instead.}

\item{refbasis}{Select the basis of the eigenspaces. See details.}

\item{maxit}{The maximum number of Newton steps allowed in empirical likelihood optimisation \insertCite{owen:2013}{TFORGE}.}

\item{evecs}{For debugging only. Supply eigenvectors of population mean.}
}
\value{
A \code{TFORGE} object including the p-value of the test (slot \code{pval}) and the statistic for \code{x} (slot \code{t0}). See \code{\link[=bootresampling]{bootresampling()}}.
}
\description{
Given a sample from a population, uses bootstrap resampling to test multiplicity hypotheses of the population mean's eigenvalues.
The test statistic is computed by \code{stat_multiplicity()}.
The null hypothesis is that the population mean has the specified the multiplicity of eigenvalues.
Bootstrap resampling is conducted from the null hypothesis, which uses the original sample converted to satisfy the null hypothesis by \code{standardise_multiplicity()}.
}
\details{
This hypothesis test works on unconstrained symmetric matrices or matrices constrained to have fixed trace.

On \code{refbasis}:
An estimate of each eigenspace specified by \code{mult} can be obtained from the eigenvectors of the sample mean.
The eigenvectors create an orthonormal basis of the (estimated) eigenspace,
however the choice of orthonormal basis for the estimated eigenspace effects the performance.
This choice is specified by the parameter \code{refbasis}.
Setting \code{refbasis = "sample"} uses the eigenvectors of the sample mean as the basis, however the resulting statistic does not appear to be pivotal.
Choosing the orthonormal basis independently of the data does result in a pivotal asymptotically chi-squared statistic.
Setting \code{refbasis = "random"} will do exactly this, by applying a uniformly random rotation of the relevant eigenvectors of the sample mean.
We recommend using \code{refbasis = "sample"} (which requires bootstrap calibration) because test power is much higher than \code{refbasis = "random"}.
We recommend that the number bootstrap resamples is at least 1000 if \code{refbasis = "sample"}.

For 3x3 matrices, the weighted-bootstrapping method used by \code{test_multiplicity_nonnegative()} has poor test size for samples smaller than 20;
larger matrices will likely need larger samples.

Due to the random rotation of the eigenvectors when \code{refbasis = "random"}, use \code{\link[=set.seed]{set.seed()}} if you want the answer to be repeatable.

\code{test_multiplicity_nonnegative()} uses weighted bootstrapping with empirical likelihood to create a population that satisfies the null hypothesis.
}
\examples{
x <- rsymm_norm(15, mean = diag(c(2, 1, 1, 0)))
test_multiplicity(x, mult = c(1, 2, 1))
}
