% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss1fixedtrace.R
\name{test_ss1fixedtrace}
\alias{test_ss1fixedtrace}
\alias{stat_ss1fixedtrace}
\title{Test eigenvalues when trace=0 and sum of square eigenvalues = 1}
\usage{
test_ss1fixedtrace(x, evals = NULL, B = 1000, maxit = 25)

stat_ss1fixedtrace(x, evals = NULL)
}
\arguments{
\item{x}{A single sample of symmetric matrices or multiple samples of symmetric matrices. See \code{\link[=as_flat]{as_flat()}}.}

\item{evals}{When \code{x} is a single sample, the null hypothesis is that the population mean has eigenvalues \code{evals}.}

\item{B}{Number of bootstrap samples.}

\item{maxit}{The maximum number of Newton steps allowed in empirical likelihood optimisation \insertCite{owen:2013}{TFORGE}.}
}
\description{
For a single sample, test eigenvalues of the population mean.
For multiple samples, test for equality of the eigenvalues of the population means.
This function is for 3x3 symmetric matrices with trace of zero and sum of squared eigenvalues of one.
These constraints combine so that the space of possible sets of (ordered) eigenvalues is 1 dimensional.
The test statistic is calculated by \code{stat_ss1fixedtrace()}.
}
\details{
The sum of squared eigenvalues constraint forces the set of eigenvalues to lie on a sphere and the trace constraint forces eigenvalues onto a plane.
Combined the constraints force eigenvalues onto a circle in 3D Euclidean space.
The test statistic accounts for these constraints by projecting eigenvalues onto a line in the plane and orthogonal to the null-hypothesis eigenvalues.
Bootstrap resampling is from an empirical distribution that satisfies the null hypothesis; for this test we use empirical likelihood \insertCite{owen:2013}{TFORGE} to find probability mass weights for each matrix in the original sample.

Eigenvalues must be distinct.
}
\section{Hypotheses}{
For a single sample the null hypothesis is that the population mean has eigenvalues of \code{evals}; the alternative hypothesis is that the eigenvalues are not \code{evals}.
For multiple samples, \code{evals} must be omitted and the null hypothesis is that the population means of each sample have the same eigenvalues.
}

\references{
\insertAllCited{}
}
