% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unconstrained_aGOE.R
\name{test_unconstrained_aGOE}
\alias{test_unconstrained_aGOE}
\title{Two Sample Test of Equal Eigenvalues Using GOE Approximation}
\usage{
test_unconstrained_aGOE(x1, x2 = NULL)
}
\arguments{
\item{x1}{A single sample of matrices (passed to \code{\link[=as_fsm]{as_fsm()}}) or
a list of two samples of matrices (passed to \code{\link[=as_kfsm]{as_kfsm()}}).}

\item{x2}{If \code{x1} is a single sample then \code{x2} must be the second sample. Otherwise \code{x2} should be \code{NULL}.}
}
\value{
A list:
\itemize{
\item \code{pval} The p-value.
\item \code{t} Value of the statistic \insertCite{@Eq. 11, @schwartzman2010gr}{TFORGE}.
\item \code{a} Plug-in estimate of the \eqn{a} in the final equation of \insertCite{@Section 2.4, @schwartzman2010gr}{TFORGE}.
\item \code{v} Plug-in estimate of the \eqn{v} in the final equation of \insertCite{@Section 2.4, @schwartzman2010gr}{TFORGE}.
\item \code{var_Lambda_evals} The variance of the eigenvalues of Schwartzman's \eqn{\Lambda}{Lambda} matrix, which may relate to the quality of the Welch-Satterthwaite approximation.
}
}
\description{
Applies the equal-eigenvalue hypothesis test between two samples by \insertCite{schwartzman2010gr;textual}{TFORGE}.
The null hypothesis is that the population means of each sample have the same eigenvalues, regardless of eigenvectors.
The test uses a statistic from the situation that both populations are Gaussian Orthogonal Ensembles (Gaussian-distributed independent elements with the variance on the off diagonal elements half that of the diagonal elements).
The distribution of this statistic for more general populations is approximated using a tangent space and the Welch-Satterthwaite approximation.
}
\details{
The test statistic is equation 11 of \insertCite{schwartzman2010gr}{TFORGE}.
The p value of the test is computed using the approximate distribution reached at the end of \insertCite{@Section 2.4, @schwartzman2010gr}{TFORGE}.
}
\references{
\insertAllCited{}
}
